refactor(md): break versionMetadata into smaller components
