+++ app/scripts/modules/core/src/managed/overview/artifact/utils.ts
@@ -6,12 +6,12 @@ import { timeDiffToString } from 'core/utils';
 
 import { MarkAsBadActionModal, PinActionModal, UnpinActionModal } from './ArtifactActionModal';
 import { ManagedWriter } from '../../ManagedWriter';
-import { VersionAction } from './VersionMetadata';
 import { ACTION_DISPLAY_NAMES, getActionStatusData } from './VersionOperation';
 import { useFetchApplicationLazyQuery } from '../../graphql/graphql-sdk';
 import { QueryArtifactVersion, QueryConstraint, QueryLifecycleStep } from '../types';
 import { OVERVIEW_VERSION_STATUSES } from '../utils';
 import { MODAL_MAX_WIDTH } from '../../utils/defaults';
+import { VersionAction } from '../../versionMetadata/MetadataComponents';
 
 export const getConstraintsStatusSummary = (constraints: QueryConstraint[]) => {
   let finalStatus: QueryConstraint['status'] = 'PASS';
