+++ app/scripts/modules/core/src/managed/versionMetadata/VersionMetadata.less
